ImageTool:
new suport to generate an aleatory name for the image.
new method called getFileFromBase64String, to create a file from a base64 string.
new method named randomName to generate an aleatory and unique name.
new method to get the concurrent application's directory.

ColorChartDetailActivity:

reference for each component of the view to add events for saving and updating of a record.